rename columns and adjust paths accordingly
